Updated to make the internals more 'lispy.'
I've introduced a lisp-like singly-linked list handler, and removed all
of the object-oriented materials.  There are still *records*, but those
are different from scoped objects in the given context.  This has really
helped me understand what's going on with the difference between the dynamic
and lexical scopes, which has been one of my biggest headaches.